update to latest vblob version

+++ drivers/fs/fs_gctmp.js
@@ -7,8 +7,22 @@ var exec = require('child_process').exec;
 
 var argv = process.argv;
 var force = false;
-for (var ii = 0; ii < argv.length; ii++)
-  if (argv[ii] === '--force') { force = true; break; } //if force, gc every file
+var gc_timestamp = null;
+for (var ii = 0; ii < argv.length; ii++) {
+  if (argv[ii] === '--force') { force = true; continue; } //if force, gc every file
+  if (argv[ii] === '--ts') {
+    if (ii+1 < argv.length) {
+      try {
+        if (isNaN(gc_timestamp=parseInt(argv[ii+1],10))) throw 'NaN';
+      } catch(err) {
+        gc_timestamp = new Date().valueOf(); //current time
+      }
+    } else {
+      gc_timestamp = new Date().valueOf(); //the time this is executed
+    }
+    continue;
+  }
+}
 var BATCH_NUM = 1;
 var root_path = argv[2];
 var PREFIX_LENGTH = 2;
@@ -38,14 +52,16 @@ buck.on('gc',function(buck_idx) {
           return;
         }
         var mtime = new Date(stats.mtime).valueOf();
-        if (current_ts < mtime + MAX_TIMEOUT && !force) {
+        if ( !force && 
+             (gc_timestamp && gc_timestamp < mtime //created later than the specified timestamp 
+               || !gc_timestamp && current_ts < mtime + MAX_TIMEOUT) ){
           evt.Counter++; evt.Batch--; //still within a valid window
           if (evt.Batch === 0) {
             evt.Batch = BATCH_NUM; evt.emit('nextbatch');
           }
           return;
         }
-        fs.unlink(fdir_path+"/"+filename,function() {} );
+        if (stats.nlink <= 2) fs.unlink(fdir_path+"/"+filename,function() {} ); //not an active version
         fs.unlink(trash_dir+"/"+filename,function() {
           evt.Counter++; evt.Batch--
           if (evt.Batch === 0) {
@@ -63,7 +79,7 @@ buck.on('gc',function(buck_idx) {
     });
     evt.emit('nextbatch');
   } catch (err) {
-    console.log(err);
+    console.error(err);
   }
 });
 for (var i = 0; i < containers.length; i++)
